requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
webrtcvad==2.0.10
numpy==1.24.3
soundfile==0.12.1
werkzeug==3.0.1
pydub==0.25.1